pyproject.toml
[project]
name = "vibehost-setup"
version = "0.1.0"
description = "One-shot provisioning script for dedicated servers"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fabric>=3.2",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "rich>=13.0",
    "pydantic>=2.0",
]

[project.scripts]
vibehost-setup = "main:main"
